+++ demo0.cpp
@@ -10,13 +10,13 @@ int main()
 	ebi b;
 	cout << "Please keyin a big number, e.g. 9876543210 or 0x24cb016ea:" << endl;
 	cin >> b;
-	cout << "Your number, b, is " << b << " (0x" << hex << b << ")" << endl;
+	cout << "Your number, b, is " << dec << b << " (0x" << hex << b << ")" << endl;
 	cout << endl;
-	cout << "a+b = " << a+b << endl;
-	cout << "a-b = " << a-b << endl;
-	cout << "a*b = " << a*b << endl;
-	cout << "a/b = " << a/b << endl;
-	cout << "a\%b = " << a%b << endl;
+	cout << "a+b = " << dec << a+b << endl;
+	cout << "a-b = " << dec << a-b << endl;
+	cout << "a*b = " << dec << a*b << endl;
+	cout << "a/b = " << dec << a/b << endl;
+	cout << "a\%b = " << dec << a%b << endl;
 
 	return 0;
 }